major refactor, now everything shares the same code, making it easier to maintain
