+++ pkg/config/lpconfig/lpconfig.go
@@ -1,36 +1,22 @@
-package config
+//go:build lp
+
+package lpconfig
 
 import (
 	"log"
+	"shlyuz/pkg/config"
 	"strconv"
 
-	"gopkg.in/ini.v1"
-
 	"shlyuz/pkg/crypto/asymmetric"
-	"shlyuz/pkg/crypto/symmetric"
 	"shlyuz/pkg/encoding/hex"
-	"shlyuz/pkg/encoding/xor"
+
+	"gopkg.in/ini.v1"
 )
 
-type ShlyuzCrypto struct {
-	PeerPk      asymmetric.PublicKey
-	SymKey      []byte
-	XorKey      int
-	CompKeyPair asymmetric.AsymmetricKeyPair
-}
-
-type ShlyuzConfig struct {
-	Id            string
-	TransportName string
-	TskChkTimer   int
-	InitSignature []byte
-	CryptoConfig  ShlyuzCrypto
-}
-
-func ParseConfig(config []byte) ShlyuzConfig {
+func ParseConfig(configToParse []byte) config.ShlyuzConfig {
 	// TODO: Parse Transport Configuration - #? TransportConfig
-	var ParsedConfig ShlyuzConfig
-	cfg, err := ini.Load(config)
+	var ParsedConfig config.ShlyuzConfig
+	cfg, err := ini.Load(configToParse)
 	if err != nil {
 		log.Fatalln("failed to load config: ", err)
 	}
@@ -64,22 +50,4 @@ func ParseConfig(config []byte) ShlyuzConfig {
 	ParsedConfig.CryptoConfig.XorKey = int(xor64Key)
 
 	return ParsedConfig
-}
-
-func ReadConfig(config []byte, xKey int, symKey []byte) symmetric.SymmetricMessage {
-	config = xor.XorMessage(config, xKey)
-	rdyConfig := symmetric.Decrypt(config, symKey)
-	if rdyConfig.IsEncrypted {
-		log.Fatalln("failed to decode config")
-	}
-	return rdyConfig
-}
-
-// Debug function for reading unencrypted configs. DO NOT USE IN PRODUCTION.
-func ReadPlaintextConfig(config []byte, symKey []byte) symmetric.SymmetricMessage {
-	var rdyConfig symmetric.SymmetricMessage
-	rdyConfig.Message = config
-	rdyConfig.IsEncrypted = false
-	rdyConfig.Key = symKey
-	return rdyConfig
 }
