+++ pkg/execution/firecollect_windows.go
@@ -1,3 +1,5 @@
+//go:build implant && (!lp || !teamserver)
+
 package firecollect
 
 import (
