+++ pkg/execution/fireforget_windows.go
@@ -1,3 +1,5 @@
+//go:build implant && (!lp || !teamserver)
+
 package fireforget
 
 import (
