+++ pkg/transaction/transaction.go
@@ -1,3 +1,5 @@
+//go:build implant && (!lp || !teamserver)
+
 package transaction
 
 import (
